FPVTL-1092: Added feature toggle

+++ src/main/java/uk/gov/hmcts/reform/prl/services/noticeofchange/NoticeOfChangePartiesService.java
@@ -665,8 +665,11 @@ public class NoticeOfChangePartiesService {
                     }
                 });
         }
-        caseAssignmentService.removeAmBarristerCaseRole(caseData,
-                                                        selectedPartyDetailsMap);
+        if (featureToggleService.isAddBarristerIsEnabled()) {
+            caseAssignmentService.removeAmBarristerCaseRole(caseData,
+                                                            selectedPartyDetailsMap);
+        }
+
         caseDataUpdated = createChangeOrgReqAndRemoveRepresentative(
             authorisation,
             caseDetails,
